+++ kororaa_graphql_api/schema/toshi_hazard/__init__.py
@@ -1,3 +1,3 @@
 from .gridded_hazard import GriddedHazard, GriddedHazardResult, RegionGridEnum, query_gridded_hazard
+from .hazard_curves import hazard_curves
 from .hazard_schema import GriddedLocation, GriddedLocationResult, ToshiHazardCurveResult
-from .toshi_hazard_rev2 import hazard_curves
