+++ kororaa_graphql_api/schema/toshi_hazard/hazard_curves.py
